requirements.txt
tensorflow==2.1.0
rasa==1.10.26
rasa-sdk==1.10.3
grakn-client==2.0.1
streamlit
altair
symspellpy
chatette